first implementation for issue #426
