+++ asammdf/blocks/mdf_v3.py
@@ -37,6 +37,7 @@ from . import v2_v3_constants as v23c
 from ..signal import Signal
 from ..version import __version__
 from .conversion_utils import conversion_transfer
+from .mdf_common import MDF_Common
 from .source_utils import Source
 from .utils import (
     as_non_byte_sized_signed_int,
@@ -72,7 +73,7 @@ logger = logging.getLogger("asammdf")
 __all__ = ["MDF3"]
 
 
-class MDF3(object):
+class MDF3(MDF_Common):
     """The *header* attibute is a *HeaderBlock*.
 
     The *groups* attribute is a list of dicts, each one with the following keys
@@ -658,112 +659,6 @@ class MDF3(object):
             )
         else:
             return vals
-
-    def _validate_channel_selection(
-        self, name=None, group=None, index=None, source=None
-    ):
-        """Gets channel comment.
-        Channel can be specified in two ways:
-
-        * using the first positional argument *name*
-
-            * if there are multiple occurrences for this channel then the
-              *group* and *index* arguments can be used to select a specific
-              group.
-            * if there are multiple occurrences for this channel and either the
-              *group* or *index* arguments is None then a warning is issued
-
-        * using the group number (keyword argument *group*) and the channel
-          number (keyword argument *index*). Use *info* method for group and
-           numbers
-
-        Parameters
-        ----------
-        name : string
-            name of channel
-        group : int
-            0-based group index
-        index : int
-            0-based channel index
-        source : str
-            can be used for multiple occurence of the same channel name to
-            filter the target channel
-
-        Returns
-        -------
-        group_index, channel_index : (int, int)
-            selected channel's group and channel index
-
-        """
-
-        if name is None:
-            if group is None or index is None:
-                message = (
-                    "Invalid arguments for channel selection: "
-                    'must give "name" or, "group" and "index"'
-                )
-                raise MdfException(message)
-            else:
-                gp_nr, ch_nr = group, index
-                if ch_nr >= 0:
-                    if gp_nr > len(self.groups) - 1:
-                        raise MdfException("Group index out of range")
-                    if index > len(self.groups[gp_nr].channels) - 1:
-                        raise MdfException("Channel index out of range")
-        else:
-            if name not in self.channels_db:
-                raise MdfException(f'Channel "{name}" not found')
-            else:
-                if source is not None:
-                    for gp_nr, ch_nr in self.channels_db[name]:
-                        source_name = self._get_source_name(gp_nr, ch_nr)
-                        if source_name == source:
-                            break
-                    else:
-                        raise MdfException(f"{name} with source {source} not found")
-                elif group is None:
-
-                    gp_nr, ch_nr = self.channels_db[name][0]
-                    if len(self.channels_db[name]) > 1:
-                        message = (
-                            f'Multiple occurances for channel "{name}". '
-                            f"Using first occurance from data group {gp_nr}. "
-                            'Provide both "group" and "index" arguments'
-                            " to select another data group"
-                        )
-                        logger.warning(message)
-
-                else:
-                    if index is not None and index < 0:
-                        gp_nr = group
-                        ch_nr = index
-                    else:
-                        for gp_nr, ch_nr in self.channels_db[name]:
-                            if gp_nr == group:
-                                if index is None:
-                                    break
-                                elif index == ch_nr:
-                                    break
-                        else:
-                            if index is None:
-                                message = f'Channel "{name}" not found in group {group}'
-                            else:
-                                message = f'Channel "{name}" not found in group {group} at index {index}'
-                            raise MdfException(message)
-
-        return gp_nr, ch_nr
-
-    def _get_source_name(self, group, index):
-        grp = self.groups[group]
-
-        if grp.channels[index].source:
-            name = grp.channels[index].source.name
-        else:
-            name = ""
-        return name
-
-    def _set_temporary_master(self, master):
-        self._master = master
 
     def _read(self, mapped=False):
         stream = self._file
@@ -2614,7 +2509,6 @@ class MDF3(object):
         data=None,
         raw=False,
         ignore_invalidation_bits=False,
-        source=None,
         record_offset=0,
         record_count=None,
     ):
@@ -2623,8 +2517,6 @@ class MDF3(object):
 
         * using the first positional argument *name*
 
-            * if *source* is given this will be first used to validate the
-              channel selection
             * if there are multiple occurances for this channel then the
               *group* and *index* arguments can be used to select a specific
               group.
@@ -2659,8 +2551,6 @@ class MDF3(object):
             `False`
         ignore_invalidation_bits : bool
             only defined to have the same API with the MDF v4
-        source : str
-            source name used to select the channel
         record_offset : int
             if *data=None* use this to select the record offset from which the
             group data should be loaded
@@ -2742,18 +2632,11 @@ class MDF3(object):
                 unit=""
                 info=None
                 comment="">
-        >>> mdf.get('Sig', source='VN7060')
-        <Signal Sig:
-                samples=[ 12.  12.  12.  12.  12.]
-                timestamps=[0 1 2 3 4]
-                unit=""
-                info=None
-                comment="">
 
         """
 
         gp_nr, ch_nr = self._validate_channel_selection(
-            name, group, index, source=source
+            name, group, index
         )
 
         original_data = data
