+++ asammdf/blocks/mdf_v4.py
@@ -56,6 +56,7 @@ from ..signal import Signal
 from ..version import __version__
 from .bus_logging_utils import extract_can_signal, extract_mux
 from .conversion_utils import conversion_transfer
+from .mdf_common import MDF_Common
 from .source_utils import Source
 from .utils import (
     all_blocks_addresses,
@@ -212,7 +213,7 @@ except:
         return offsets[:-1], offsets[-1]
 
 
-class MDF4(object):
+class MDF4(MDF_Common):
     """The *header* attibute is a *HeaderBlock*.
 
     The *groups* attribute is a list of dicts, each one with the following keys:
@@ -1658,119 +1659,6 @@ class MDF4(object):
 
         return parents, dtypes
 
-    @lru_cache(maxsize=1024)
-    def _validate_channel_selection(
-        self, name=None, group=None, index=None, source=None
-    ):
-        """Gets channel comment.
-        Channel can be specified in two ways:
-
-        * using the first positional argument *name*
-
-            * if there are multiple occurrences for this channel then the
-            *group* and *index* arguments can be used to select a specific
-            group.
-            * if there are multiple occurrences for this channel and either the
-            *group* or *index* arguments is None then a warning is issued
-
-        * using the group number (keyword argument *group*) and the channel
-        number (keyword argument *index*). Use *info* method for group and
-        channel numbers
-
-
-        Parameters
-        ----------
-        name : string
-            name of channel
-        group : int
-            0-based group index
-        index : int
-            0-based channel index
-        source : str
-            can be used for multiple occurence of the same channel name to
-            filter the target channel
-
-        Returns
-        -------
-        group_index, channel_index : (int, int)
-            selected channel's group and channel index
-
-        """
-
-        if name is None:
-            if group is None or index is None:
-                message = (
-                    "Invalid arguments for channel selection: "
-                    'must give "name" or, "group" and "index"'
-                )
-                raise MdfException(message)
-            else:
-                gp_nr, ch_nr = group, index
-                if ch_nr >= 0:
-                    try:
-                        grp = self.groups[gp_nr]
-                    except IndexError:
-                        raise MdfException("Group index out of range")
-
-                    try:
-                        grp.channels[ch_nr]
-                    except IndexError:
-                        raise MdfException(
-                            f"Channel index out of range: {(name, group, index)}"
-                        )
-        else:
-            if name not in self.channels_db:
-                raise MdfException(f'Channel "{name}" not found')
-            else:
-                if source is not None:
-                    for gp_nr, ch_nr in self.channels_db[name]:
-                        if source in self._get_source_name(gp_nr, ch_nr):
-                            break
-                    else:
-                        raise MdfException(f"{name} with source {source} not found")
-                elif group is None:
-
-                    gp_nr, ch_nr = self.channels_db[name][0]
-                    if len(self.channels_db[name]) > 1:
-                        message = (
-                            f'Multiple occurances for channel "{name}". '
-                            f"Using first occurance from data group {gp_nr}. "
-                            'Provide both "group" and "index" arguments'
-                            " to select another data group"
-                        )
-                        logger.warning(message)
-
-                else:
-                    if index is not None and index < 0:
-                        gp_nr = group
-                        ch_nr = index
-                    else:
-                        for gp_nr, ch_nr in self.channels_db[name]:
-                            if gp_nr == group:
-                                if index is None:
-                                    break
-                                elif index == ch_nr:
-                                    break
-                        else:
-                            if index is None:
-                                message = f'Channel "{name}" not found in group {group}'
-                            else:
-                                message = f'Channel "{name}" not found in group {group} at index {index}'
-                            raise MdfException(message)
-
-        return gp_nr, ch_nr
-
-    def _get_source_name(self, group, index):
-        source = self.groups[group].channels[index].source
-        cn_source = source.name if source else ""
-
-        source = self.groups[group].channel_group.acq_source
-        cg_source = source.name if source else ""
-
-        return (cn_source, cg_source)
-
-    def _set_temporary_master(self, master):
-        self._master = master
 
     def _get_data_blocks_info(
         self,
@@ -5871,7 +5759,6 @@ class MDF4(object):
         data=None,
         raw=False,
         ignore_invalidation_bits=False,
-        source=None,
         record_offset=0,
         record_count=None,
     ):
@@ -5883,8 +5770,6 @@ class MDF4(object):
 
         * using the first positional argument *name*
 
-            * if *source* is given this will be first used to validate the
-              channel selection
             * if there are multiple occurances for this channel then the
               *group* and *index* arguments can be used to select a specific
               group.
@@ -5918,8 +5803,6 @@ class MDF4(object):
             `False`
         ignore_invalidation_bits : bool
             option to ignore invalidation bits
-        source : str
-            source name used to select the channel
         record_offset : int
             if *data=None* use this to select the record offset from which the
             group data should be loaded
@@ -6007,20 +5890,11 @@ class MDF4(object):
                 unit=""
                 info=None
                 comment="">
-        >>> # validation using source name
-        ...
-        >>> mdf.get('Sig', source='VN7060')
-        <Signal Sig:
-                samples=[ 12.  12.  12.  12.  12.]
-                timestamps=[0 1 2 3 4]
-                unit=""
-                info=None
-                comment="">
 
         """
 
         gp_nr, ch_nr = self._validate_channel_selection(
-            name, group, index, source=source
+            name, group, index
         )
 
         grp = self.groups[gp_nr]
